Organization of methods inside do_utils and dbdo_utils so that all methods who act directly on the database are inside dbdo_utils and methods that do not are inside do_utils
